suzerain: #1888 add --Pa/--Pb restart tests for -np 2
git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@22311 0bd85f0b-c151-43b9-a1e9-27231104551f

+++ apps/test_channel_setup.sh
@@ -27,17 +27,17 @@ if test x$prereq_status != x ; then
     exit $prereq_status
 fi
 
-# Create temporary directory and clean it up on exit
+# Create temporary directory and remove it on exit (unless TEST_CHANNEL_DEBUG is set)
 tmpdir=`mktemp -d`
-trap "rm -rf $tmpdir" EXIT
+test -z "${TEST_CHANNEL_DEBUG-}" && trap "rm -rf $tmpdir" EXIT
 
 # Minimalistic command execution infrastructure
 banner_prefix=`basename $0`
 banner() { echo; echo $banner_prefix${METACASE+ (}${METACASE-}${METACASE+)}: "$@" ; }
-run()    { echo mpiexec -np 1        "$@" ; mpiexec -np 1        "$@"            ; }
-runq()   { echo mpiexec -np 1        "$@" ; mpiexec -np 1        "$@" >/dev/null ; }
-prun()   { echo mpiexec -np ${NP:-1} "$@" ; mpiexec -np ${NP:-1} "$@"            ; }
-prunq()  { echo mpiexec -np ${NP:-1} "$@" ; mpiexec -np ${NP:-1} "$@" >/dev/null ; }
+run()    { echo mpiexec -np 1        "$@" ; mpiexec -np 1        "$@"             ; }
+runq()   { echo mpiexec -np 1        "$@" ; mpiexec -np 1        "$@" > /dev/null ; }
+prun()   { echo mpiexec -np ${NP:-1} "$@" ; mpiexec -np ${NP:-1} "$@"             ; }
+prunq()  { echo mpiexec -np ${NP:-1} "$@" ; mpiexec -np ${NP:-1} "$@" > /dev/null ; }
 differ() { echo h5diff "$@" ; h5diff "$@" || h5diff -rv "$@" ;}
 
 banner "Creating initial field to use for tests"
@@ -48,3 +48,4 @@ declare -ir htdelta=1
 declare -ir Nz=6
 runq ./channel_init "$tmpdir/mms0.h5" --mms=0                            \
                     --Nx=$Nx --Ny=$Ny --k=$k --htdelta=$htdelta --Nz=$Nz
+chmod +r "$tmpdir/mms0.h5"
